api/requirements.txt
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.6.0
matplotlib
numpy